% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{MAMSNews}
\alias{MAMSNews}
\title{Shows changes and news}
\usage{
MAMSNews()
}
\value{
Screen output.
}
\description{
Functions showing changes since previous versions.
}
\details{
Displays the changes and news given in the NEWS file of the package.
}
\examples{
\dontrun{MAMSNews()}
}
\author{
Thomas Jaki
}
