% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneous.R
\name{print.MAMS.stepdown}
\alias{print.MAMS.stepdown}
\title{Generic print function for class MAMS.stepdown.}
\usage{
\method{print}{MAMS.stepdown}(x, digits = max(3, getOption("digits") - 4), ...)
}
\arguments{
\item{x}{An output object of class MAMS}

\item{digits}{Number of significant digits to be printed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Text output.
}
\description{
Generic print function for class MAMS.stepdown.
}
\details{
print produces a brief summary of an object from class MAMS.stepdown
including boundaries and requires sample size if initially requested.
}
