% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneous.R
\name{summary.MAMS.stepdown}
\alias{summary.MAMS.stepdown}
\title{Generic summary function for class MAMS.stepdown.}
\usage{
\method{summary}{MAMS.stepdown}(object, digits = max(3, getOption("digits") - 4), ...)
}
\arguments{
\item{object}{An output object of class MAMS}

\item{digits}{Number of significant digits to be printed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Text output.
}
\description{
Generic summary function for class MAMS.stepdown.
}
\details{
print produces a brief summary of an object from class MAMS.stepdown
including boundaries and requires sample size if initially requested.
}
